--- experiments/CRP2_CPE/fit.py
@@ -1,3 +1,0 @@
-
-def exp_0():
-    pass
