--- experiments/CRP2_CPE/params.json
@@ -1,71 +1,0 @@
-{
-    "REQUIRED_PARAMETERS": ["rA", "rB", "rX", "KAA", "KAB", "KBA", "KBB"],
-
-    "PARAMETER_GROUPS": {
-        "IRREVERSIBLE": ["p0", "p1", "p2"],
-        "REVERSIBLE": ["p3", "p4", "p5"]
-    },
-
-    "PARAMETER_SETS": {
-        "p0": {
-            "rA": 1.0,
-            "rB": 1.0,
-            "rX": 1.0,
-            "KAA": 0.0,
-            "KAB": 0.0,
-            "KBA": 0.0,
-            "KBB": 0.0
-        },
-
-        "p1": {
-            "rA": 2.0,
-            "rB": 2.0,
-            "rX": 1.0,
-            "KAA": 0.0,
-            "KAB": 0.0,
-            "KBA": 0.0,
-            "KBB": 0.0
-        },
-
-        "p2": {
-            "rA": 3.0,
-            "rB": 3.0,
-            "rX": 1.0,
-            "KAA": 0.0,
-            "KAB": 0.0,
-            "KBA": 0.0,
-            "KBB": 0.0
-        },
-
-        "p3": {
-            "rA": 10.0,
-            "rB": 0.5,
-            "rX": 1.0,
-            "KAA": 0.0,
-            "KAB": 0.0,
-            "KBA": 0.0,
-            "KBB": 0.0
-        },
-
-        "p4": {
-            "rA": 4.0,
-            "rB": 2.2,
-            "rX": 1.0,
-            "KAA": 0.0,
-            "KAB": 0.0,
-            "KBA": 0.0,
-            "KBB": 0.0
-        },
-
-        "p5": {
-            "rA": 5.0,
-            "rB": 2.5,
-            "rX": 1.0,
-            "KAA": 0.2,
-            "KAB": 0.0,
-            "KBA": 0.4,
-            "KBB": 0.0
-        }
-    }
-    
-}
