--- experiments/LV/params.json
@@ -1,46 +1,0 @@
-{
-  "REQUIRED_PARAMETERS": ["alpha", "beta", "gamma", "delta"],
-  "PARAMETER_GROUPS": {
-    "Mammals": ["Rabbit-Fox", "Deer-Wolf"],
-    "Marine": ["Fish-Shark", "Plankton-Fish"],
-    "Insects": ["Insect-Bird", "Aphid-Ladybug"]
-  },
-  "PARAMETER_SETS": {
-    "Rabbit-Fox": {
-      "alpha": 2.0,
-      "beta": 0.4,
-      "delta": 0.4,
-      "gamma": 0.3
-    },
-    "Deer-Wolf": {
-      "alpha": 0.7,
-      "beta": 0.3,
-      "delta": 0.2,
-      "gamma": 0.2
-    },
-    "Fish-Shark": {
-      "alpha": 1.5,
-      "beta": 0.5,
-      "delta": 0.3,
-      "gamma": 0.4
-    },
-    "Plankton-Fish": {
-      "alpha": 2.5,
-      "beta": 0.6,
-      "delta": 0.5,
-      "gamma": 0.5
-    },
-    "Insect-Bird": {
-      "alpha": 3.0,
-      "beta": 0.8,
-      "delta": 0.3,
-      "gamma": 0.6
-    },
-    "Aphid-Ladybug": {
-      "alpha": 2.8,
-      "beta": 0.7,
-      "delta": 0.4,
-      "gamma": 0.5
-    }
-  }
-}
