+++ polypesto/models/__init__.py
@@ -1,0 +1,3 @@
+from . import sbml_util as sbml
+
+__all__ = ["sbml"]
