--- polypesto/models/sbml/__init__.py
@@ -1,36 +1,0 @@
-
-from .util import (
-    write_model,
-    create_model,
-    create_compartment,
-    create_species,
-    create_parameter,
-    create_initial_assignment,
-    create_reaction,
-    create_rule,
-    create_rate_rule,
-    create_algebraic_rule,
-    add_termination_event,
-    validateSBML,
-    ModelDefinition,
-    Model,
-    Document
-)
-
-__all__ = [
-    "write_model",
-    "create_model",
-    "create_compartment",
-    "create_species",
-    "create_parameter",
-    "create_initial_assignment",
-    "create_reaction",
-    "create_rule",
-    "create_rate_rule",
-    "create_algebraic_rule",
-    "add_termination_event",
-    "validateSBML",
-    "ModelDefinition",
-    "Model",
-    "Document"
-]
